Correction of bug
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@6959 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ examples/simdag/dax/dax_test.c
@@ -42,7 +42,7 @@ int main(int argc, char **argv) {
   if (argc == 3) {
     char *last=strrchr(argv[2],'.');
 
-    tracefilename=bprintf("%.*s.trace",last==NULL?strlen(argv[2]):last-argv[2],argv[2]);
+    tracefilename=bprintf("%.*s.trace",(int)(last==NULL?strlen(argv[2]):last-argv[2]),argv[2]);
   } else {
     tracefilename = xbt_strdup(argv[3]);
   }
